gobelo_poster/requirements.txt
flask>=3.0
pyyaml>=6.0
waitress
